First setup
Change ports for network, button + for new port. Docker container IP.
