+++ packages/admin-ui/src/pages/packages/components/Network/IpsByService.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function IpsByService({ ip }: { ip?: string }) {
+  return (
+    <>
+      <p>Container IP: {ip ? ip : "Not available"}</p>
+    </>
+  );
+}
