+++ packages/admin-ui/src/pages/packages/components/Network/network.scss
@@ -1,4 +1,4 @@
-.ports-editor {
+.network-editor {
   table {
     width: 100%;
